pyproject.toml
[project]
name = "cascade"
version = "0.0.1"
authors = [
    { name = "Logan Ward", email = "lward@anl.gov" },
    { name = "Logan Ward", email = "lward@anl.gov" },
]
description = 'Fast and reliable dynamic simulations by concurrently learning, auditing, and using surrogates'
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
keywords = ["HPC", "AI", "Workflows", "Quantum Chemistry", "Chemical Engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: System :: Distributed Computing",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "ase>3.22",
]

[tool.setuptools.packages.find]
include = ['cascade*']

[project.optional-dependencies]
test = [
    'nbstripout',  # Used as a pre-commit hook
    'flake8',
    'pytest',
    'pytest-forked',
    'pytest-timeout',
    'pytest-cov',
]
ani = [
    'torchani',
    'h5py'
]
